fix opengl viewport scaling bug on mac
qt's opengl contexts reportedly do not take display scaling into consideration, so we have to apply the scaling ourselves to get context-appropriate pixel dimensions.  this is observed on retina laptops as the opengl context only rendering in the bottom left quadrant (as it needs a 2x scaling in X and Y).