Migrated from Qt5.14.2 to Qt6.4.2
The problem with Display is that it generates a conflicting declaration error
with typedef struct _XDisplay Display present in QtGui. The better way to fix
this issue is obviously to wrap all Arbalest code inside a namespace Arbalest,
but this might not be worth the effort, at least for now.