Added functionaly to drag MainWindow while it is maximized
Dragging the MainWindow title bar while it is maximized is now possible.
Also, now it isn't possible to drag the MainWindow around while a menu
is open, it isn't possible to drag the MainWindow around after the title
bar was grabbed from one of its buttons, and it is possible to move the
MainWindow around grabbing the top left Arbalest icon.